refactor: revert event bus logic
