+++ src/event/eventBus.ts
@@ -3,6 +3,8 @@ import GridEvent from './gridEvent';
 import { getInstance } from '../instance';
 import { GridId } from '../store/types';
 
+const eventBusMap: { [id: number]: EventBus } = {};
+
 export interface EventBus {
   on: Function;
   off: Function;
@@ -11,7 +13,7 @@ export interface EventBus {
 
 export function createEventBus(id: GridId) {
   const listenersMap: { [eventName: string]: Function[] } = {};
-  return {
+  eventBusMap[id] = {
     on(eventName: string, func: Function) {
       const listeners = listenersMap[eventName];
       listenersMap[eventName] = listeners ? [...listeners, func] : [func];
@@ -38,4 +40,10 @@ export function createEventBus(id: GridId) {
       }
     }
   };
+
+  return eventBusMap[id];
+}
+
+export function getEventBus(id: GridId) {
+  return eventBusMap[id];
 }
